+++ src/store/index.ts
@@ -15,6 +15,8 @@ import { simulationSettingsApi } from "./simulationSettingsApi";
 import { simulationSettingsReducer } from "./simulationSettingsSlice";
 import { tabReducer } from "./tabSlice";
 import { geometryIssueReducer } from "./geometryIssueSlice";
+import { materialReducer } from "./materialSlice";
+import { userPreferencesApi } from "./userPreferenceApi";
 
 export const store = configureStore({
   reducer: {
@@ -24,6 +26,7 @@ export const store = configureStore({
     [materialsApi.reducerPath]: materialsApi.reducer,
     [simulationSettingsApi.reducerPath]: simulationSettingsApi.reducer,
     [auralizationApi.reducerPath]: auralizationApi.reducer,
+    [userPreferencesApi.reducerPath]: userPreferencesApi.reducer,
     project: projectReducer,
     simulation: simulationReducer,
     model: modelReducer,
@@ -33,6 +36,7 @@ export const store = configureStore({
     simulationSettings: simulationSettingsReducer,
     tab: tabReducer,
     geometryIssue: geometryIssueReducer,
+    material: materialReducer,
   },
 
   middleware: (getDefaultMiddleware) =>
@@ -59,6 +63,7 @@ export const store = configureStore({
       materialsApi.middleware,
       simulationSettingsApi.middleware,
       auralizationApi.middleware,
+      userPreferencesApi.middleware,
     ),
 });